+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.3.2
+
+- Fix bug
+- Add createAdvancedQuoteMessage method for MessageManager
+- Add createAdvancedTextMessage method for MessageManager
+
 ## 2.3.0+1
 
 - Fix bug
@@ -189,33 +195,42 @@ Fix bug
 2. Add createFileMessageFromFullPath method
 
 ## 0.0.9+4
+
 Fix ios friendship params error
 
 ## 0.0.9+3
+
 Ios native remove method 'removeAdvancedMsgListener'
 
 ## 0.0.9+2
+
 1. Fix bug
 2. Upgrade sdk
 
 ## 0.0.9+1
+
 Fix bug
 
 ## 0.0.9
+
 Add ios support
 
 ## 0.0.8
+
 Upgrade sdk to 0.0.14 </br>
 Sending a message adds a retry mechanism
 
 ## 0.0.7
+
 1. Add a new method to obtain conversation id based on user id or group id </br>
 2. Modify the type of conversation manager input parameter sessionType to int
 
 ## 0.0.6
+
 Fix bug
 
 ## 0.0.5
+
 Upgrade sdk </br>
 Fix message error bug
 
@@ -224,11 +239,14 @@ Fix message error bug
 Replace the input parameters of the createCustomMessage method from Uint8List to Map
 
 ## 0.0.3
+
 Upgrade sdk </br>
 Fix bug </br>
 
 ## 0.0.2
-1. Move the markSingleMessageHasRead, markGroupMessageHasRead and getTotalUnreadMsgCount methods from the MessageManager class to the ConversationManager class. </br>
+
+1. Move the markSingleMessageHasRead, markGroupMessageHasRead and getTotalUnreadMsgCount methods
+   from the MessageManager class to the ConversationManager class. </br>
 2. Modify the parameters of the setFriendInfo method. </br>
 3. Modify the parameters of the createGroup and setGroupInfo methods.